pyproject.toml
[project]
name = "travel-planner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "langchain>=1.2.12",
    "langchain-groq>=1.1.2",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.1.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.41.0",
]
